// src/schedule/dto/Lesson.dto.ts
export interface LessonDto {
  auditorium: string;
  auditoriumAmount: number;
  auditoriumGUID: string;
  auditoriumOid: number;
  author: string;
  beginLesson: string;
  building: string;
  buildingGid: number;
  buildingOid: number;
  contentOfLoadOid: number;
  contentOfLoadUID: any | null;
  contentTableOfLessonsName: string;
  contentTableOfLessonsOid: number;
  createddate: string;
  date: string;
  dateOfNest: string;
  dayOfWeek: number;
  dayOfWeekString: string;
  detailInfo: string;
  discipline: string;
  disciplineOid: number;
  disciplineinplan: any | null;
  disciplinetypeload: number;
  duration: number;
  endLesson: string;
  group: any | null;
  groupGUID: any | null;
  groupOid: number;
  groupUID: any | null;
  group_facultyname: any | null;
  group_facultyoid: number;
  hideincapacity: number;
  isBan: false;
  kindOfWork: string;
  kindOfWorkComplexity: number;
  kindOfWorkOid: number;
  kindOfWorkUid: any | null;
  lecturer: string;
  lecturerCustomUID: any | null;
  lecturerEmail: any | null;
  lecturerGUID: string;
  lecturerOid: number;
  lecturerUID: any | null;
  lecturer_post_oid: number;
  lecturer_rank: string;
  lecturer_title: string;
  lessonNumberEnd: number;
  lessonNumberStart: number;
  lessonOid: number;
  listGroups: any[];
  listOfLecturers: {
    lecturer: string;
    lecturerCustomUID: any | null;
    lecturerEmail: any | null;
    lecturerGUID: any | null;
    lecturerOid: number;
    lecturerUID: any | null;
    lecturer_post_oid: number;
    lecturer_rank: string;
    lecturer_title: string;
  }[];
  modifieddate: string;
  note: any | null;
  note_description: string;
  parentSchedule_Status: number;
  parentschedule: string;
  replaces: any | null;
  specialization_name: any | null;
  specialization_oid: number;
  stream: string;
  streamOid: number;
  stream_facultyoid: number;
  subGroup: any | null;
  subGroupOid: number;
  subgroup_facultyoid: number;
  subgroup_groupOid: number;
  tableofLessonsName: string;
  tableofLessonsOid: number;
  url1: string;
  url1_description: string;
  url2: string;
  url2_description: string;
}